requirements.txt
PyYAML>=6.0,<7.0
pandas>=2.0,<3.0
openpyxl>=3.1,<4.0
customtkinter>=5.2,<6.0
Pillow>=10.0,<13.0
